fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyJWT # For decoding JWTs
python-dotenv # For reading the .env file
pymongo